backend/requirements.txt
contractions==0.1.73
Django==4.1.3
djangorestframework==3.14.0
environ==1.0
gnews==0.2.7
GoogleNews==1.6.7
matplotlib==3.7.1
newsapi==0.1.1
nltk==3.8.1
numpy==1.24.2
pandas==2.0.0
pydantic==1.8.2
rake_nltk==1.0.6
scikit_learn==1.2.2
textblob==0.17.1
wordcloud==1.8.2.2
newsapi-python
django-filter
django-cors-headers
python-environ
news-fetch